issue solved

+++ booking_management/lib/vendor_side/model/vendor_subscription_plan_screen_models/get_all_subscription_model.dart
@@ -140,7 +140,7 @@ class SubscriptionWorkerList {
   String name;
   String detail;
   String interval;
-  int price;
+  double price;
   String image;
   bool isActive;
   String currency;
@@ -155,7 +155,7 @@ class SubscriptionWorkerList {
         name: json["name"] ?? "",
         detail: json["detail"] ?? "",
         interval: json["interval"] ?? "",
-        price: json["price"].toInt() ?? 0,
+        price: json["price"] ?? 0.0,
         image: json["image"] ?? "",
         isActive: json["isActive"] ?? false,
         currency: json["currency"] ?? "",
